DataExporter: Enhance DataExporter to support other components #4414
