+++ src/main/java/org/primefaces/component/export/DataExporter.java
@@ -133,13 +133,12 @@ public class DataExporter implements ActionListener, StateHolder {
                 }
                 else {
                     UIComponent component = (UIComponent) components.get(0);
-                    if (!(component instanceof DataTable)) {
+                    if (!(component instanceof DataTable) && this.customExporter == null) {
                         throw new FacesException("Unsupported datasource target:\"" + component.getClass().getName()
-                                + "\", exporter must target a PrimeFaces DataTable.");
+                                + "\", exporter must target a PrimeFaces DataTable or provide a customExporter for " + component.getClass().getName());
                     }
 
-                    DataTable table = (DataTable) component;
-                    exporter.export(context, table, outputFileName, isPageOnly, isSelectionOnly, encodingType,
+                    exporter.export(context, component, outputFileName, isPageOnly, isSelectionOnly, encodingType,
                             preProcessor, postProcessor, exporterOptions, onTableRender);
                 }
             }
